+++ examples/resources/nxos_cli/resource.tf
@@ -1,0 +1,6 @@
+resource "nxos_cli" "example" {
+  cli = <<-EOT
+  interface loopback123
+  description configured-via-cli
+  EOT
+}