+++ internal/provider/provider.go
@@ -334,6 +334,7 @@ func (p *NxosProvider) Resources(ctx context.Context) []func() resource.Resource
 	return []func() resource.Resource{
 		NewRestResource,
 		NewSaveConfigResource,
+		NewCliResource,
 		NewBGPResource,
 		NewBGPAddressFamilyResource,
 		NewBGPAdvertisedPrefixResource,